fixes #148: example of MUC invite contains escaped characters.
This was a result of trying to make the same entity usable for the pre-exising 'invite' endpoint (that works with just one JID), as well as for the new 'invite' endpoint (that can use a collection of JIDs). This introduced a list-of-JID field to the entity, which obviously should not be used for the singular-type endpoint.

In this commit, I've introduced distinct entities to be used for each endpoint, to clear up any confusion. The old entity was reverted to its original state. The new entity inherits from the old endpoint, and adds the 'list-of-jids-to-invite' field.
