+++ changelog.html
@@ -46,6 +46,7 @@ REST API Plugin Changelog
 
 <p><b>1.9.1</b> (tbd)</p>
 <ul>
+    <li>[<a href='https://github.com/igniterealtime/openfire-restAPI-plugin/issues/148'>#148</a>] - Example of MUC invite contains escaped characters</li>
     <li>[<a href='https://github.com/igniterealtime/openfire-restAPI-plugin/issues/146'>#146</a>] - Chatroom 'affiliation' URL template clash</li>
 </ul>
