+++ src/java/org/jivesoftware/openfire/plugin/rest/entity/MUCInvitationEntity.java
@@ -16,21 +16,15 @@
 
 package org.jivesoftware.openfire.plugin.rest.entity;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlElementWrapper;
 import javax.xml.bind.annotation.XmlRootElement;
-import java.util.ArrayList;
-import java.util.List;
 
 @XmlRootElement(name = "mucInvitation")
 public class MUCInvitationEntity {
 
     String reason;
-
-    private List<String> jidsToInvite;
 
     public MUCInvitationEntity() {
     }
@@ -43,20 +37,5 @@ public class MUCInvitationEntity {
 
     public void setReason(String reason) {
         this.reason = reason;
-    }
-
-    @XmlElementWrapper(name = "jidsToInvite")
-    @XmlElement(name = "jid")
-    @JsonProperty(value = "jidsToInvite")
-    @Schema(description = "The JIDs and/or names of the users and groups to invite into the room", example = "<jidsToInvite><jid>jane@example.org</jid></jidsToInvite>")
-    public List<String> getJidsToInvite() {
-        if (jidsToInvite == null) {
-            jidsToInvite = new ArrayList<>();
-        }
-        return jidsToInvite;
-    }
-
-    public void setJidsToInvite(List<String> jidsToInvite) {
-        this.jidsToInvite = jidsToInvite;
     }
 }
